Separated goLinkData into its own file
